Introduce effective way to decide the Stats data source
