+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/ui/mystore/MyStoreViewModel.kt
@@ -4,7 +4,11 @@ import android.os.Parcelable
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.asLiveData
 import androidx.navigation.NavHostController
+import com.woocommerce.android.phone.PhoneConnectionRepository
+import com.woocommerce.android.system.NetworkStatus
 import com.woocommerce.android.ui.login.LoginRepository
+import com.woocommerce.android.ui.mystore.datasource.FetchStatsFromPhone
+import com.woocommerce.android.ui.mystore.datasource.FetchStatsFromStore
 import com.woocommerce.commons.viewmodel.ScopedViewModel
 import com.woocommerce.commons.viewmodel.getStateFlow
 import dagger.assisted.Assisted
@@ -22,7 +26,10 @@ import org.wordpress.android.fluxc.model.SiteModel
 @Suppress("UnusedPrivateProperty")
 @HiltViewModel(assistedFactory = MyStoreViewModel.Factory::class)
 class MyStoreViewModel @AssistedInject constructor(
+    private val phoneRepository: PhoneConnectionRepository,
     private val fetchStatsFromStore: FetchStatsFromStore,
+    private val fetchStatsFromPhone: FetchStatsFromPhone,
+    private val networkStatus: NetworkStatus,
     @Assisted private val navController: NavHostController,
     loginRepository: LoginRepository,
     savedState: SavedStateHandle
@@ -42,10 +49,16 @@ class MyStoreViewModel @AssistedInject constructor(
             }.launchIn(this)
     }
 
+    private suspend fun evaluateStatsSource(selectedSite: SiteModel) = when {
+        networkStatus.isConnected() -> fetchStatsFromStore(selectedSite)
+        phoneRepository.isPhoneConnectionAvailable() -> fetchStatsFromPhone()
+        else -> throw IllegalStateException("No connection available")
+    }
+
     private fun requestStoreStats(selectedSite: SiteModel) {
         _viewState.update { it.copy(isLoading = true) }
         launch {
-            fetchStatsFromStore(selectedSite)
+            evaluateStatsSource(selectedSite)
                 .onEach { statsData ->
                     _viewState.update {
                         it.copy(
