+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/ui/mystore/datasource/MyStoreStatsData.kt
@@ -1,0 +1,32 @@
+package com.woocommerce.android.ui.mystore.datasource
+
+import com.woocommerce.commons.extensions.convertedFrom
+
+data class RevenueData(
+    val totalRevenue: Double,
+    val orderCount: Int
+)
+
+data class MyStoreStatsData(
+    private val revenueData: RevenueData?,
+    private val visitorData: Int?
+) {
+    val isFinished
+        get() = revenueData != null &&
+            visitorData != null
+    val revenue
+        get() = revenueData?.totalRevenue ?: 0.0
+
+    val ordersCount
+        get() = revenueData?.orderCount ?: 0
+
+    val visitorsCount
+        get() = visitorData ?: 0
+
+    val conversionRate: String
+        get() {
+            val ordersCount = revenueData?.orderCount ?: 0
+            val visitorsCount = visitorData ?: 0
+            return ordersCount convertedFrom visitorsCount
+        }
+}
